Initial implementation of the faa_array_queue
This is a very simple implementation of a fast lock-free concurrent
queue. It now seems to work and behave as expected, but is not that well
scrutinized yet when it comes to correctness or efficiency. However, it
does perform similarly, or a bit better, than LCRQ.

It is now completely generic, but you could make it much faster if only
handling raw u64.