Restructure repo as workspace with fewer macros
This changes the repo structure from a single package to a workspace of
many packages, where the benchmarks are in one package, and each data
structrue in its own. This ensures each data structure is compiled
separately as before, while relying way less on obfuscating feature
switches. It also exchanges a core benchmark macro with a function,
making it easier to work with.

This also separates the logic of everything in a nicer way, reducing
complexity when getting into the project.